requirements.txt
asgiref==3.7.1
Django==4.2.1
django-colorfield==0.9.0
django-cors-headers==4.1.0
djangorestframework==3.14.0
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
